Refactor: 커뮤니티 페이지 구조 변경 cst -> ssr
